refactor: architecture improvements and bug fixes
- Addressed silent failure patterns in storage and pg-notification-listener by adding explicit error logging.
- Improved type safety in ai-service by utilizing native type narrowing and type guards over unsafe assertions.
- Extracted EditorToolbar and QueryResults components from the monolithic tab-query-editor, reducing its size by over 1,000 lines.
